Output a more useful resolve error
This outputs a missage that blames a specific sife for not having DNSLink record.  For example, the error message looks like:

`could not resolve name: bad.example.net is missing DNSLink record (https://docs.ipfs.io/concepts/dnslink/)`

The "could not resolve name" portion is still present because the returned error wraps the original ErrResolveFailed, allowing code to test if the error is an ErrorResolveFailed error.