feat: add *Many vectorised publishing operations
Thoses operations do the same thing than the scalar operations they replace but accepts many elements, this allows to use the PutMany interface on routings that supports it and yield a subsentantial speed boost.

All the scalar function that have been replaced has been replaced by stubs that just move to the vectorized codepath ASAP.
This is done to increase maintainability of the code and is, to me, an acceptable tradeoff given that this is cold code runned once each 12 hours.

That increase maintainability because if the scalar functions kept a scalar codepath, it would be faster for single publishes, but that requires us to maintain two codebase, any change to the scalar codebase would need to be done to the vectorised one too without the compiler warning you of the requirement and vice versa.